Job Restructure Finished
