+++ src/Commands/Sync.js
@@ -13,7 +13,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 const dna_discord_framework_1 = require("dna-discord-framework");
 const OrcaBotDataManager_1 = __importDefault(require("../OrcaBotDataManager"));
-const OrcaJobManager_1 = __importDefault(require("../OrcaJobManager"));
+const OrcaJobManager_1 = __importDefault(require("../OrcaJob/OrcaJobManager"));
 /**
  * Command that
  */
